UXD-805 docs(ListBox): cleanup file names and remove files
