+++ packages/ListBox/stories/examples/Multi/CustomCheckboxes.js
@@ -1,22 +1,31 @@
 import React from "react";
+import Checkbox from "@paprika/checkbox";
 import * as characters from "../../fixtures/characters";
 import ListBox from "../../../src";
 
-function OptionCustomCheckbox(props) {
-  return (
-    <span>
-      <input tabIndex="-1" type="checkbox" checked={props.isSelected} />
-      {props.character}
-    </span>
-  );
-}
+export const CustomCheckboxes = () => {
+  const { CHECKED, UNCHECKED } = Checkbox.types.state;
 
-export const CustomCheckboxes = () => (
-  <ListBox isMulti>
-    {characters.villians.map(Option => (
-      <ListBox.Option label={Option.props.children} key={Option.props.children}>
-        {({ isSelected }) => <OptionCustomCheckbox isSelected={isSelected} character={Option.props.children} />}
-      </ListBox.Option>
-    ))}
-  </ListBox>
-);
+  const [checkedState, setCheckedState] = React.useState(UNCHECKED);
+
+  const handleChange = () => setCheckedState(checkedState === CHECKED ? UNCHECKED : CHECKED);
+
+  return (
+    <ListBox isMulti>
+      {characters.villians.map(Option => (
+        <ListBox.Option label={Option.props.children} key={Option.props.children}>
+          {({ isSelected }) => (
+            <div style={{ display: "flex", alignItems: "center" }}>
+              <Checkbox
+                checkedState={isSelected ? CHECKED : checkedState}
+                onChange={handleChange}
+                style={{ paddingRight: "8px" }}
+              />
+              {Option.props.children}
+            </div>
+          )}
+        </ListBox.Option>
+      ))}
+    </ListBox>
+  );
+};
